Added stat tracker functionality
Changed table thickness and color in the style.css

Gave functionality to stat tracker, allowing table to be updated with their stats (for now, will be added to table through button with JavaScript function on the page)

Added function to generate the table on the stat tracker page. No persistency yet, mainly functionality proof-of-concept.